# [libc++] Fold __search_substring into _Traits::find in case the second string has length 1