… libcxx/test/std/strings/basic.string/string.ops/string_find/string_size.pass.cpp